+++ packages/react/src/components/Loader/Loader.tsx
@@ -14,6 +14,7 @@ import { loaderBehavior } from '../../lib/accessibility'
 import { Accessibility } from '../../lib/accessibility/types'
 import { WithAsProp, ShorthandValue, withSafeTypeForAs } from '../../types'
 import Box, { BoxProps } from '../Box/Box'
+import Text, { TextProps } from '../Text/Text'
 
 export interface LoaderSlotClassNames {
   indicator: string
@@ -35,7 +36,7 @@ export interface LoaderProps extends UIComponentProps, ColorComponentProps {
   inline?: boolean
 
   /** A loader can contain a label. */
-  label?: ShorthandValue<BoxProps>
+  label?: ShorthandValue<TextProps>
 
   /** A label in the loader can have different positions. */
   labelPosition?: 'above' | 'below' | 'start' | 'end'
@@ -134,7 +135,7 @@ class Loader extends UIComponent<WithAsProp<LoaderProps>, LoaderState> {
               styles: styles.indicator,
             },
           })}
-          {Box.create(label, {
+          {Text.create(label, {
             defaultProps: { className: Loader.slotClassNames.label, styles: styles.label },
           })}
         </ElementType>